% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_socket_clients.R
\name{close_socket_clients}
\alias{close_socket_clients}
\alias{closeSocketClients}
\title{Close one or more clients currently connected}
\usage{
close_socket_clients(sockets = "all", server_port = 8888)

closeSocketClients(sockets = "all", server_port = 8888)
}
\arguments{
\item{sockets}{the list of socket client names (sockXXX) to close, or \code{"all"}
(by default) to disconnect all currently connected clients.}

\item{server_port}{the corresponding R socket server port.}
}
\description{
The socket servers asks to clients to nicely disconnect (possibly doing
further process on their side). This function is used by
\code{\link[=stop_socket_server]{stop_socket_server()}}, but it can also be invoked manually to ask for
disconnection of a particular client. Note that, in this case, the client
still can decide not to disconnect! The code send to ask for client
disconnection is: \verb{\\\\f}.
}
\seealso{
\code{\link[=send_socket_clients]{send_socket_clients()}}
}
\concept{stateful socket server interprocess communication}
\keyword{IO}
