% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_socket.R
\name{get_socket_servers}
\alias{get_socket_servers}
\alias{getSocketServers}
\title{Get the ports of current R socket servers}
\usage{
get_socket_servers()

getSocketServers()
}
\value{
A character string vector, or \code{NULL} if no R socket server is currently
running.
}
\description{
Returns a list with all the ports of currently running R socket servers.
}
\seealso{
\code{\link[=get_socket_clients]{get_socket_clients()}}, \code{\link[=get_socket_server_name]{get_socket_server_name()}}, \link{start_socket_server}
}
\concept{stateful socket server interprocess communication}
\keyword{IO}
\keyword{utilities}
