% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIncompleteLines.R
\name{supportsIncompleteLines}
\alias{supportsIncompleteLines}
\title{Is used to indicate that this is a modified version of the original package}
\usage{
supportsIncompleteLines()
}
\description{
Is used to indicate that this is a modified version of the original package
}
