% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSocket.R
\name{getSocketServers}
\alias{getSocketServers}
\title{Get the ports of current R socket servers}
\usage{
getSocketServers()
}
\value{
A character string vector, or \code{NULL} if no R socket server is currently
running.
}
\description{
Returns a list with all the ports of currently running R socket servers.
}
\seealso{
\code{\link[=getSocketClients]{getSocketClients()}}, \code{\link[=getSocketServerName]{getSocketServerName()}}, \link{startSocketServer}
}
\concept{stateful socket server interprocess communication}
\keyword{IO}
\keyword{utilities}
